Add a cleanup option
The cleanup option can be used to find versions that are
present in the database but not in the main tree. All those
versions will be
  a) logged, and
  b) deleted from the database

This way, outdated data can be found, logged and cleaned
up. This option will currently be used to find outdated data,
that is bugs in the updater logic. Once these bugs are found
and fix this option becomes obsolete.

Signed-off-by: Max Magorsch <arzano@gentoo.org>