Integrate information from external services
This is a major revision of the application. That is the application
can display:
  - repology.org data
  - qa-reports (pkgcheck)
  - pull-requests
  - bugs
  - security bugs
  - dependencies
for each package now. Apart from that a page for each maintainer
(project, developer or proxied maintainer) has been added, that
contains information about packages he maintains, as:
  - outdated packages
  - pull-requests
  - bugs
  - security bugs
  - changelog
Finally the design has also been slightly altered to display all of
the new information.

Signed-off-by: Max Magorsch <arzano@gentoo.org>