Add mailto link to the maintainer metadata again
In former times the maintainer metadata included a
mailto link including the email address of the
maintainer. This link has recently been replaced
by a link to the overview page of the maintainer.

Now a small envelope icon has been added additional
to the existing link to the maintainer overview page.
The envelope icon contains a mailto link as in former
times.

Signed-off-by: Max Magorsch <arzano@gentoo.org>