Improve the ignored repositories for repology
So far one had to specify all versions of a repository
to ignore all of them, e.g.:

  app-forensics/afl::rosa_2016_1
  app-forensics/afl::rosa_2019_1

Now it's possible to use

  app-forensics/afl::rosa

to ignore all versions of the repository. However,
it's still possible to use e.g. rosa_2019_1 to just
ignore rosa_2019_1.

Thanks to hanno for pointing this out.

Signed-off-by: Max Magorsch <arzano@gentoo.org>