Add a periodical selfcheck
This can be used to check whether any of the information
in the database are outdated. The findings are exported
as prometheus metric.

In former times this check has been done externally. Now
it is directly integrated into soko as lately there have
been some reports about outdated information again.

Please note that this is currently still a WIP and rather
a first approach to get a feeling for the current situation
then a perfect solution.

Signed-off-by: Max Magorsch <arzano@gentoo.org>