bin: make updates more reliable and faster
`git pull --rebase` has many corner cases where it tries to do a "safe"
update and may leave the working tree in a unsuitable state for
automated processes.

Since the working git repo is meant to be entirely automated, split it
out to a git-fetch of just the branch we need, and then forcibly reset
the tree content to the tip of that branch after each fetch.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>