Exclude commits from repo mirror bots from parsing
Since we switched from repo/gentoo.org to repo/sync/gentoo.org we
got a lot of new commits by different repo mirror bots. At the time
of writting there are / have been at least two of them. Commits of
these bots are not desired to be shown in the changelog (as they are
also not part of repo/gentoo.org).

For now they are just removed by excluding all commits with a specific,
hardcoded author email address. This works for the moment, as we just have
commits of two different bots.

In future this should be replaced by a better solution, which is not using
hardcoded emails. For now this is used to fix the wrong data in packages.g.o
in production though.

Signed-off-by: Max Magorsch <arzano@gentoo.org>