index: fix DoS attack based on SQL Injection
If the user enables another view of the index page, using the "Recently
Visited Packages" page (in user preferences), the index page uses the
value of the `search_history` cookie as base64-encoded comma-separated
list of atoms. Those atoms are string loaded directly into the SQL query
with `atom = '%s'` format string. This allows the user to inject SQL
queries into the query, which can be used to DoS the server (for example
truncate all the tables, inject erroneous data, etc). In some time I'll
open a full report of this solved vulnerability.

Fix it by using parameterized queries with `pg.In` function, which is
safer and faster.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>