updater/bugs: fetch old bugs
The previous implementation was using the search facility with csv
output, which had hard limit of 10000. No passing of offset & limit
enabled us to fetch the next pages. By using the REST API, we can fetch
beyond that limit without issues.

Also implement a better "diff" update, which will only fetch the changed
since a week ago. This will reduce the load on the server, and will
enable us to update bugs more frequently. Reuse the `last_commit` field
in DB to hold the last update time, which enables us to start with
"empty" DB and still get the full history on next deploy.

Closes: https://bugs.gentoo.org/768174
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>