app/packages/changelog: super optimize the changes query
A truly addition of a dragon SQL query. This query filters the changed
files (added, deleted, modified) on DB side, fetches only relevant,
reconstructs a type that matches ORM, and builds a changelog around it.

Huge performance improvement, especially on packages which has the huge
tree-wide commits, huge page size reduction and more.

Sadly, I can't make it more readable, because it's already a miracle I'm
able to perform all this logic in SQL, but this is very much worth it.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>