packages: auto redirect for old names of pkgmoved packages
Parse and store all the package moves from the portage tree. Done every
time we perform small delta update (and those files are updated), and
also on full update.

Then when accessing a package page, if the package wasn't found, we
attempt to verify if it existed as a pkgmoved package, and if so,
perform permanent redirect to the new package page, while attempting
to preserve the query parameters and sub path.

Closes: https://bugs.gentoo.org/946666
Resolves: https://github.com/gentoo/soko/issues/33
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>